Light refactoring.
- Make public API clearer.
- Move base test case classes to sub-folder and rename as harnesses.
