+++ Sources/MySqlConnector/MySqlClient.swift
@@ -12,9 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import BinaryCodable
 import Foundation
-import Socket
 
 /**
  An instance of MySqlClient lazily creates and maintains connections to a single MySql server.
@@ -29,146 +27,122 @@ public final class MySqlClient {
    - parameter password: The password to authenticate with. Can be empty if no password is required.
    - parameter database: If provided, this database will be the default used database for queries.
    */
-  public init(to host: String, port: Int32 = 3306, username: String, password: String, database: String? = nil) {
+  public init(to host: String, port: Int32 = 3306, username: String, password: String?, database: String? = nil) {
     self.host = host
     self.port = port
     self.username = username
-    self.password = password
+    self.password = password ?? ""
     self.database = database
   }
 
-  private var connectionPool: [Connection] = []
-  private let host: String
-  private let port: Int32
-  private let username: String
-  private let password: String
-  private let database: String?
+  /**
+   Sends a MySql query to the MySql server and returns the server's response, parsing any returned rows using the
+   provided Decodable `rowType`.
+
+   - parameter rowType: A Decodable type that will be used to decode the row's response values.
+   - parameter query: A single MySql query statement to be performed on the MySql server.
+   - returns: `.ERR` if query MySql resulted in an error on the MySql server. `.OK` if the query did not return any rows
+     but was otherwise successful. `.Results` if the query was successful and returned rows.
+   - throws: `ClientError.noConnectionAvailable` if no connection could be made with the server.
+   */
+  @discardableResult
+  public func query<T: Decodable>(_ query: String, rowType type: T.Type) throws -> QueryResponse<T> {
+    guard let connection = try anyIdleConnection() else {
+      throw ClientError.noConnectionAvailable
+    }
+
+    do {
+      try connection.send(payload: Query(query))
+
+      connection.isIdle = false
+
+      let response = try connection.read(payloadType: GenericResponse.self)
+
+      switch response {
+      case .ERR(let errorCode, let errorMessage):
+        connection.isIdle = true
+        return .ERR(errorCode: errorCode,
+                    errorMessage: errorMessage)
+
+      case .OK(let context):
+        connection.isIdle = true
+        return .OK(numberOfAffectedRows: context.numberOfAffectedRows,
+                   lastInsertId: context.lastInsertId,
+                   info: context.info)
+
+      case .ResultSetColumnCount(let columnCount):
+        let iterator = try QueryResultDecoder<T>(columnCount: columnCount,
+                                                 connection: connection)
+        return .Results(iterator: AnyIterator {
+          let next = iterator.next()
+          if next == nil {
+            connection.isIdle = true
+          }
+          return next
+        })
+
+      default:
+        preconditionFailure("Unimplemented.")
+      }
+    } catch let error {
+      connection.isIdle = true
+      throw error
+    }
+  }
+
+  /**
+   A Decodable type that can be used to parse returned MySql rows as generic dictionaries.
+   */
+  public typealias DictionaryRow = [String: String?]
+
+  /**
+   Sends a MySql query to the MySql server and returns the server's response, parsing any returned rows as a dictionary.
+
+   - parameter query: A single MySql query statement to be performed on the MySql server.
+   - returns: `.ERR` if query MySql resulted in an error on the MySql server. `.OK` if the query did not return any rows
+   but was otherwise successful. `.Results` if the query was successful and returned rows.
+   - throws: `ClientError.noConnectionAvailable` if no connection could be made with the server.
+   */
+  @discardableResult
+  public func query(_ query: String) throws -> QueryResponse<DictionaryRow> {
+    return try self.query(query, rowType: DictionaryRow.self)
+  }
+  
+  var connectionPool: [Connection] = []
+  let host: String
+  let port: Int32
+  let username: String
+  let password: String
+  let database: String?
 }
 
 /**
- An error that occurs when connecting to a MySql server.
+ A representation of the response from a MySqlClient query.
  */
-public enum ClientError: Error, Equatable {
+public enum QueryResponse<T: Decodable> {
   /**
-   Indicates that no connection could be made to the MySql server. Check the host and port.
+   Indicates that the query succeeded.
+
+   - parameter numberOfAffectedRows: The number of rows affected by the query.
+   - parameter lastInsertId: The id of the last-inserted row.
+   - parameter info: A human-readable status information related to the query.
    */
-  case noConnectionAvailable
+  case OK(numberOfAffectedRows: UInt64, lastInsertId: UInt64, info: String?)
 
   /**
-   Indicates that an error was received by the MySql server when attempting to perform the handshake.
+   Indicates that the query failed.
+
+   - parameter errorCode: The error code returned by the MySql server.
+   - parameter errorMessage: A human-readable error message returned by the MySql server.
    */
-  case handshakeError(errorCode: ErrorCode)
-}
+  case ERR(errorCode: ErrorCode, errorMessage: String)
 
-private let clientCapabilityFlags: CapabilityFlags = [
-  .connectWithDb,
-  .deprecateEof,
-  .protocol41,
-  .secureConnection,
-  .sessionTrack,
-  .pluginAuth,
-  .multiStatements,
-  .multiResults
-]
+  /**
+   Indicates that the query succeeded with a lazily-iterable list of rows.
 
-extension MySqlClient {
-  func anyIdleConnection() throws -> Connection? {
-    if let connection = connectionPool.first(where: { $0.isIdle }) {
-      return connection
-    }
-    guard let connection = try MySqlClient.connect(to: host, port: port, username: username, password: password, database: database) else {
-      return nil
-    }
-    connectionPool.append(connection)
-    return connection
-  }
+   The provided iterator will only pull data from the server as it is enumerated.
 
-  private static func connect(to host: String,
-                              port: Int32 = 3306,
-                              username: String,
-                              password: String,
-                              database: String? = nil) throws -> Connection? {
-    let socket = try Socket.create()
-    try socket.connect(to: host, port: port)
-    guard socket.isConnected else {
-      return nil
-    }
-
-    let socketDataStream = createDataStream(socket: socket)
-    var decoder = BinaryDataDecoder()
-
-    // Step 1: Receive Handshake from server.
-    let handshake = try decoder.decode(Packet<Handshake>.self, from: socketDataStream)
-
-    // > The client should only announce the capabilities in the Handshake
-    // > Response Packet that it has in common with the server.
-    // https://dev.mysql.com/doc/internals/en/capability-negotiation.html
-    let capabilityFlags = clientCapabilityFlags.intersection(handshake.payload.serverCapabilityFlags)
-
-    // Subsequent decodings on this data stream may require access to the intersected capability flags, so we make them
-    // available through `userInfo`.
-    decoder.userInfo[.capabilityFlags] = capabilityFlags
-
-    // Step 2: Send HandshakeResponse to server.
-    let handshakeResponse = try HandshakeResponse(username: username,
-                                                  password: password,
-                                                  database: database,
-                                                  capabilityFlags: capabilityFlags,
-                                                  authPluginData: handshake.payload.authPluginData,
-                                                  authPluginName: handshake.payload.authPluginName)
-      .encodedAsPacket(sequenceNumber: 1)
-    try socket.write(from: handshakeResponse)
-
-    // Step 3: Confirm that an OK response was received.
-    let response = try decoder.decode(Packet<GenericResponse>.self, from: socketDataStream)
-
-    if case .OK = response.payload {
-      return Connection(decoder: decoder, socket: socket, socketDataStream: socketDataStream)
-    } else if case .ERR(let errorCode, _) = response.payload {
-      throw ClientError.handshakeError(errorCode: errorCode)
-    }
-    return nil
-  }
-
-  private static func createDataStream(socket: Socket) -> BufferedData {
-    var buffer = Data(capacity: socket.readBufferSize)
-    return BufferedData(reader: AnyBufferedDataSource(read: { length in
-      if buffer.count == 0 {
-        _ = try socket.read(into: &buffer)
-      }
-      let pulledData = buffer.prefix(length)
-      buffer = buffer.dropFirst(length)
-      return pulledData
-    }, isAtEnd: {
-      do {
-        return try buffer.isEmpty && !socket.isReadableOrWritable(waitForever: false, timeout: 0).readable
-      } catch {
-        return true
-      }
-    }))
-  }
-}
-
-final class Connection {
-  let decoder: BinaryDataDecoder
-  let socket: Socket
-  let socketDataStream: BufferedData
-
-  init(decoder: BinaryDataDecoder, socket: Socket, socketDataStream: BufferedData) {
-    self.decoder = decoder
-    self.socket = socket
-    self.socketDataStream = socketDataStream
-  }
-
-  // Indicates whether or not this connection is able to send data.
-  var isIdle = true
-
-  func send<Payload: BinaryEncodable>(payload: Payload) throws {
-    precondition(isIdle, "A connection is being reused even though it is not yet idle.")
-    try socket.write(from: payload.encodedAsPacket(sequenceNumber: 0))
-  }
-
-  func read<Payload: BinaryDecodable>(payloadType: Payload.Type) throws -> Payload {
-    return try decoder.decode(Packet<Payload>.self, from: socketDataStream).payload
-  }
+   - parameter iterator: A lazy iterator of the query's rows.
+   */
+  case Results(iterator: AnyIterator<T>)
 }
